\docType{methods}
\name{flowDaTest}
\alias{flowDaTest}
\alias{flowDaTest,flowDa-method}
\title{Create slot for testing differences between treatments in the discriminant space}
\arguments{
  \item{disc}{the discriminants for which testing has to be
  performed}

  \item{nPerm}{the number of permutations that has to be
  performed for establishing the permutation distributions
  of differences in the discriminant space}

  \item{what="p.value"}{for establishing the null
  distribution of the p-values and "what=statistic" for
  establishing the null distribtion of the statistic.}
}
\description{
  Create slot for testing differences between treatments in
  the discriminant space
}
\references{
  De Roy, K., Clement, L., Thas, O., Wang, Y., and Boon, N.
  (2012). Flow cytometry for fast microbial community
  fingerprinting. Water Research, 46 (3), 907-919.
}

