\name{flowFDA-package}
\alias{flowFDA-package}
\alias{flowFDA}
\docType{package}
\title{
flowFDA: flow cytometry Functional Data Analysis
}
\description{
This package provides functions for analysing flow cytometry experiments
    with model based clustering, functional principal component analysis,
    functional discriminant analysis and to compare multivariate flow cytometry
    fingerprints accross treatments

\tabular{ll}{
Package: \tab flowFDA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-21\cr
License: \tab GPL-2\cr
}
}

